requirements.txt
Faker==40.12.0
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2026.1
